requirements.txt
streamlit
groq
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pydantic
python-dotenv
